refactor core/event substructure
re-add some things for event.py restructuring that got lost in rebase to master
