+++ misc/docs/source/packages/obspy.core.event.rst
@@ -1,0 +1,38 @@
+.. currentmodule:: obspy.core.inventory
+.. automodule:: obspy.core.inventory
+
+    .. comment to end block
+
+    Classes & Functions
+    -------------------
+    .. autosummary::
+       :toctree: autogen
+       :nosignatures:
+
+       catalog.Catalog
+       event.Event
+       origin.Origin
+       magnitude.Magnitude
+       origin.Pick
+       magnitude.Amplitude
+       source.FocalMechanism
+       source.MomentTensor
+
+    .. comment to end block
+
+    Modules
+    -------
+    .. autosummary::
+       :toctree: autogen
+       :nosignatures:
+
+       base
+       catalog
+       event
+       header
+       magnitude
+       origin
+       radpattern
+       source
+
+    .. comment to end block
