+++ obspy/core/event/base.py
@@ -28,25 +28,15 @@ import collections
 import copy
 import inspect
 import re
-import sys
 import warnings
 import weakref
 from copy import deepcopy
 from uuid import uuid4
 
-from obspy.core.event_header import (AmplitudeCategory, AmplitudeUnit,
-                                     DataUsedWaveType, EvaluationMode,
-                                     EvaluationStatus, EventDescriptionType,
-                                     MomentTensorCategory, MTInversionType,
-                                     OriginDepthType, OriginType,
-                                     OriginUncertaintyDescription, PickOnset,
-                                     PickPolarity, SourceTimeFunctionType)
+from obspy.core.event.header import DataUsedWaveType, ATTRIBUTE_HAS_ERRORS
 from obspy.core.utcdatetime import UTCDateTime
 from obspy.core.util import AttribDict
 from obspy.core.util.decorator import deprecated
-
-
-ATTRIBUTE_HAS_ERRORS = True
 
 
 class QuantityError(AttribDict):
@@ -1068,441 +1058,6 @@ class WaveformStreamID(__WaveformStreamID):
             self.channel_code if self.channel_code else "")
 
 
-__Amplitude = _event_type_class_factory(
-    "__Amplitude",
-    class_attributes=[("resource_id", ResourceIdentifier),
-                      ("generic_amplitude", float, ATTRIBUTE_HAS_ERRORS),
-                      ("type", str),
-                      ("category", AmplitudeCategory),
-                      ("unit", AmplitudeUnit),
-                      ("method_id", ResourceIdentifier),
-                      ("period", float, ATTRIBUTE_HAS_ERRORS),
-                      ("snr", float),
-                      ("time_window", TimeWindow),
-                      ("pick_id", ResourceIdentifier),
-                      ("waveform_id", WaveformStreamID),
-                      ("filter_id", ResourceIdentifier),
-                      ("scaling_time", UTCDateTime, ATTRIBUTE_HAS_ERRORS),
-                      ("magnitude_hint", str),
-                      ("evaluation_mode", EvaluationMode),
-                      ("evaluation_status", EvaluationStatus),
-                      ("creation_info", CreationInfo)],
-    class_contains=["comments"])
-
-
-class Amplitude(__Amplitude):
-    """
-    This class represents a quantification of the waveform anomaly, usually a
-    single amplitude measurement or a measurement of the visible signal
-    duration for duration magnitudes.
-
-    :type resource_id: :class:`~obspy.core.event.ResourceIdentifier`
-    :param resource_id: Resource identifier of Amplitude.
-    :type force_resource_id: bool, optional
-    :param force_resource_id: If set to False, the automatic initialization of
-        `resource_id` attribute in case it is not specified will be skipped.
-    :type generic_amplitude: float
-    :param generic_amplitude: Measured amplitude value for the given
-        waveformID. Note that this attribute can describe different physical
-        quantities, depending on the type and category of the amplitude. These
-        can be, e.g., displacement, velocity, or a period. If the only
-        amplitude information is a period, it has to specified here, not in the
-        period attribute. The latter can be used if the amplitude measurement
-        contains information on, e.g., displacement and an additional period.
-        Since the physical quantity described by this attribute is not fixed,
-        the unit of measurement cannot be defined in advance. However, the
-        quantity has to be specified in SI base units. The enumeration given in
-        attribute unit provides the most likely units that could be needed
-        here. For clarity, using the optional unit attribute is highly
-        encouraged.
-    :type generic_amplitude_errors:
-        :class:`~obspy.core.util.attribdict.AttribDict`
-    :param generic_amplitude_errors: AttribDict containing error quantities.
-    :type type: str, optional
-    :param type: Describes the type of amplitude using the nomenclature from
-        Storchak et al. (2003). Possible values are:
-
-        * unspecified amplitude reading (``'A'``),
-        * amplitude reading for local magnitude (``'AML'``),
-        * amplitude reading for body wave magnitude (``'AMB'``),
-        * amplitude reading for surface wave magnitude (``'AMS'``), and
-        * time of visible end of record for duration magnitude (``'END'``).
-
-    :type category: str, optional
-    :param category:  Amplitude category.  This attribute describes the way the
-        waveform trace is evaluated to derive an amplitude value. This can be
-        just reading a single value for a given point in time (point), taking a
-        mean value over a time interval (mean), integrating the trace over a
-        time interval (integral), specifying just a time interval (duration),
-        or evaluating a period (period).
-        Possible values are:
-
-        * ``"point"``,
-        * ``"mean"``,
-        * ``"duration"``,
-        * ``"period"``,
-        * ``"integral"``,
-        * ``"other"``
-
-    :type unit: str, optional
-    :param unit: Amplitude unit. This attribute provides the most likely
-        measurement units for the physical quantity described in the
-        genericAmplitude attribute. Possible values are specified as
-        combinations of SI base units.
-        Possible values are:
-
-        * ``"m"``,
-        * ``"s"``,
-        * ``"m/s"``,
-        * ``"m/(s*s)"``,
-        * ``"m*s"``,
-        * ``"dimensionless"``,
-        * ``"other"``
-
-    :type method_id: :class:`~obspy.core.event.ResourceIdentifier`, optional
-    :param method_id: Describes the method of amplitude determination.
-    :type period: float, optional
-    :param period: Dominant period in the timeWindow in case of amplitude
-        measurements. Not used for duration magnitude.  Unit: s
-    :type snr: float, optional
-    :param snr: Signal-to-noise ratio of the spectrogram at the location the
-        amplitude was measured.
-    :type time_window: :class:`~obspy.core.event.TimeWindow`, optional
-    :param time_window: Description of the time window used for amplitude
-        measurement. Recommended for duration magnitudes.
-    :type pick_id: :class:`~obspy.core.event.ResourceIdentifier`, optional
-    :param pick_id: Refers to the ``resource_id`` of an associated
-        :class:`~obspy.core.event.Pick` object.
-    :type waveform_id: :class:`~obspy.core.event.ResourceIdentifier`, optional
-    :param waveform_id: Identifies the waveform stream on which the amplitude
-        was measured.
-    :type filter_id: :class:`~obspy.core.event.ResourceIdentifier`, optional
-    :param filter_id: Identifies the filter or filter setup used for filtering
-        the waveform stream referenced by ``waveform_id``.
-    :type scaling_time: :class:`~obspy.core.utcdatetime.UTCDateTime`, optional
-    :param scaling_time: Scaling time for amplitude measurement.
-    :type scaling_time_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param scaling_time_errors: AttribDict containing error quantities.
-    :type magnitude_hint: str, optional
-    :param magnitude_hint: Type of magnitude the amplitude measurement is used
-        for.  This is a free-text field because it is impossible to cover all
-        existing magnitude type designations with an enumeration. Possible
-        values are:
-
-        * unspecified magnitude (``'M'``),
-        * local magnitude (``'ML'``),
-        * body wave magnitude (``'Mb'``),
-        * surface wave magnitude (``'MS'``),
-        * moment magnitude (``'Mw'``),
-        * duration magnitude (``'Md'``)
-        * coda magnitude (``'Mc'``)
-        * ``'MH'``, ``'Mwp'``, ``'M50'``, ``'M100'``, etc.
-
-    :type evaluation_mode: str, optional
-    :param evaluation_mode: Evaluation mode of Amplitude. Allowed values are
-        the following:
-
-        * ``"manual"``
-        * ``"automatic"``
-
-    :type evaluation_status: str, optional
-    :param evaluation_status: Evaluation status of Amplitude. Allowed values
-        are the following:
-
-        * ``"preliminary"``
-        * ``"confirmed"``
-        * ``"reviewed"``
-        * ``"final"``
-        * ``"rejected"``
-        * ``"reported"``
-
-    :type comments: list of :class:`~obspy.core.event.Comment`, optional
-    :param comments: Additional comments.
-    :type creation_info: :class:`~obspy.core.event.CreationInfo`, optional
-    :param creation_info: CreationInfo for the Amplitude object.
-
-    .. note::
-
-        For handling additional information not covered by the QuakeML
-        standard and how to output it to QuakeML see the
-        :ref:`ObsPy Tutorial <quakeml-extra>`.
-    """
-
-
-__Pick = _event_type_class_factory(
-    "__Pick",
-    class_attributes=[("resource_id", ResourceIdentifier),
-                      ("time", UTCDateTime, ATTRIBUTE_HAS_ERRORS),
-                      ("waveform_id", WaveformStreamID),
-                      ("filter_id", ResourceIdentifier),
-                      ("method_id", ResourceIdentifier),
-                      ("horizontal_slowness", float, ATTRIBUTE_HAS_ERRORS),
-                      ("backazimuth", float, ATTRIBUTE_HAS_ERRORS),
-                      ("slowness_method_id", ResourceIdentifier),
-                      ("onset", PickOnset),
-                      ("phase_hint", str),
-                      ("polarity", PickPolarity),
-                      ("evaluation_mode", EvaluationMode),
-                      ("evaluation_status", EvaluationStatus),
-                      ("creation_info", CreationInfo)],
-    class_contains=["comments"])
-
-
-class Pick(__Pick):
-    """
-    A pick is the observation of an amplitude anomaly in a seismogram at a
-    specific point in time. It is not necessarily related to a seismic event.
-
-    :type resource_id: :class:`~obspy.core.event.ResourceIdentifier`
-    :param resource_id: Resource identifier of Pick.
-    :type force_resource_id: bool, optional
-    :param force_resource_id: If set to False, the automatic initialization of
-        `resource_id` attribute in case it is not specified will be skipped.
-    :type time: :class:`~obspy.core.utcdatetime.UTCDateTime`
-    :param time: Observed onset time of signal (“pick time”).
-    :type time_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param time_errors: AttribDict containing error quantities.
-    :type waveform_id: :class:`~obspy.core.event.WaveformStreamID`
-    :param waveform_id: Identifies the waveform stream.
-    :type filter_id: :class:`~obspy.core.event.ResourceIdentifier`, optional
-    :param filter_id: Identifies the filter or filter setup used for filtering
-        the waveform stream referenced by waveform_id.
-    :type method_id: :class:`~obspy.core.event.ResourceIdentifier`, optional
-    :param method_id: Identifies the picker that produced the pick. This can be
-        either a detection software program or a person.
-    :type horizontal_slowness: float, optional
-    :param horizontal_slowness: Observed horizontal slowness of the signal.
-        Most relevant in array measurements. Unit: s·deg^(−1)
-    :type horizontal_slowness_errors:
-        :class:`~obspy.core.util.attribdict.AttribDict`
-    :param horizontal_slowness_errors: AttribDict containing error quantities.
-    :type backazimuth: float, optional
-    :param backazimuth: Observed backazimuth of the signal. Most relevant in
-        array measurements. Unit: deg
-    :type backazimuth_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param backazimuth_errors: AttribDict containing error quantities.
-    :type slowness_method_id: :class:`~obspy.core.event.ResourceIdentifier`,
-        optional
-    :param slowness_method_id: Identifies the method that was used to determine
-        the slowness.
-    :type onset: str, optional
-    :param onset: Flag that roughly categorizes the sharpness of the onset.
-        Allowed values are:
-
-        * ``"emergent"``
-        * ``"impulsive"``
-        * ``"questionable"``
-
-    :type phase_hint: str, optional
-    :param phase_hint: Tentative phase identification as specified by the
-        picker.
-    :type polarity: str, optional
-    :param polarity: Indicates the polarity of first motion, usually from
-        impulsive onsets. Allowed values are:
-
-        * ``"positive"``
-        * ``"negative"``
-        * ``"undecidable"``
-
-    :type evaluation_mode: str, optional
-    :param evaluation_mode: Evaluation mode of Pick. Allowed values are the
-        following:
-
-        * ``"manual"``
-        * ``"automatic"``
-
-    :type evaluation_status: str, optional
-    :param evaluation_status: Evaluation status of Pick. Allowed values are
-        the following:
-
-        * ``"preliminary"``
-        * ``"confirmed"``
-        * ``"reviewed"``
-        * ``"final"``
-        * ``"rejected"``
-        * ``"reported"``
-
-    :type comments: list of :class:`~obspy.core.event.Comment`, optional
-    :param comments: Additional comments.
-    :type creation_info: :class:`~obspy.core.event.CreationInfo`, optional
-    :param creation_info: CreationInfo for the Pick object.
-
-    .. note::
-
-        For handling additional information not covered by the QuakeML
-        standard and how to output it to QuakeML see the
-        :ref:`ObsPy Tutorial <quakeml-extra>`.
-    """
-
-
-__Arrival = _event_type_class_factory(
-    "__Arrival",
-    class_attributes=[("resource_id", ResourceIdentifier),
-                      ("pick_id", ResourceIdentifier),
-                      ("phase", str),
-                      ("time_correction", float),
-                      ("azimuth", float),
-                      ("distance", float),
-                      ("takeoff_angle", float, ATTRIBUTE_HAS_ERRORS),
-                      ("time_residual", float),
-                      ("horizontal_slowness_residual", float),
-                      ("backazimuth_residual", float),
-                      ("time_weight", float),
-                      ("horizontal_slowness_weight", float),
-                      ("backazimuth_weight", float),
-                      ("earth_model_id", ResourceIdentifier),
-                      ("creation_info", CreationInfo)],
-    class_contains=["comments"])
-
-
-class Arrival(__Arrival):
-    """
-    Successful association of a pick with an origin qualifies this pick as an
-    arrival. An arrival thus connects a pick with an origin and provides
-    additional attributes that describe this relationship. Usually
-    qualification of a pick as an arrival for a given origin is a hypothesis,
-    which is based on assumptions about the type of arrival (phase) as well as
-    observed and (on the basis of an earth model) computed arrival times, or
-    the residual, respectively. Additional pick attributes like the horizontal
-    slowness and backazimuth of the observed wave—especially if derived from
-    array data—may further constrain the nature of the arrival.
-
-    :type resource_id: :class:`~obspy.core.event.ResourceIdentifier`
-    :param resource_id: Resource identifier of Arrival.
-    :type force_resource_id: bool, optional
-    :param force_resource_id: If set to False, the automatic initialization of
-        `resource_id` attribute in case it is not specified will be skipped.
-    :type pick_id: :class:`~obspy.core.event.ResourceIdentifier`
-    :param pick_id: Refers to the resource_id of a Pick.
-    :type phase: str
-    :param phase: Phase identification. For possible values, please refer to
-        the description of the Phase object.
-    :type time_correction: float, optional
-    :param time_correction: Time correction value. Usually, a value
-        characteristic for the station at which the pick was detected,
-        sometimes also characteristic for the phase type or the slowness.
-        Unit: s
-    :type azimuth: float, optional
-    :param azimuth: Azimuth of station as seen from the epicenter. Unit: deg
-    :type distance: float, optional
-    :param distance: Epicentral distance. Unit: deg
-    :type takeoff_angle: float, optional
-    :param takeoff_angle: Angle of emerging ray at the source, measured against
-        the downward normal direction. Unit: deg
-    :type takeoff_angle_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param takeoff_angle_errors: AttribDict containing error quantities.
-    :type time_residual: float, optional
-    :param time_residual: Residual between observed and expected arrival time
-        assuming proper phase identification and given the earth_model_ID of
-        the Origin, taking into account the time_correction. Unit: s
-    :type horizontal_slowness_residual: float, optional
-    :param horizontal_slowness_residual: Residual of horizontal slowness and
-        the expected slowness given the current origin (refers to attribute
-        horizontal_slowness of class Pick).
-    :type backazimuth_residual: float, optional
-    :param backazimuth_residual: Residual of backazimuth and the backazimuth
-        computed for the current origin (refers to attribute backazimuth of
-        class Pick).
-    :type time_weight: float, optional
-    :param time_weight: Weight of the arrival time for computation of the
-        associated Origin. Note that the sum of all weights is not required to
-        be unity.
-    :type horizontal_slowness_weight: float, optional
-    :param horizontal_slowness_weight: Weight of the horizontal slowness for
-        computation of the associated Origin. Note that the sum of all weights
-        is not required to be unity.
-    :type backazimuth_weight: float, optional
-    :param backazimuth_weight: Weight of the backazimuth for computation of the
-        associated Origin. Note that the sum of all weights is not required to
-        be unity.
-    :type earth_model_id: :class:`~obspy.core.event.ResourceIdentifier`,
-        optional
-    :param earth_model_id: Earth model which is used for the association of
-        Arrival to Pick and computation of the residuals.
-    :type comments: list of :class:`~obspy.core.event.Comment`, optional
-    :param comments: Additional comments.
-    :type creation_info: :class:`~obspy.core.event.CreationInfo`, optional
-    :param creation_info: CreationInfo for the Arrival object.
-
-    .. note::
-
-        For handling additional information not covered by the QuakeML
-        standard and how to output it to QuakeML see the
-        :ref:`ObsPy Tutorial <quakeml-extra>`.
-    """
-
-
-__OriginQuality = _event_type_class_factory(
-    "__OriginQuality",
-    class_attributes=[("associated_phase_count", int),
-                      ("used_phase_count", int),
-                      ("associated_station_count", int),
-                      ("used_station_count", int),
-                      ("depth_phase_count", int),
-                      ("standard_error", float),
-                      ("azimuthal_gap", float),
-                      ("secondary_azimuthal_gap", float),
-                      ("ground_truth_level", str),
-                      ("minimum_distance", float),
-                      ("maximum_distance", float),
-                      ("median_distance", float)])
-
-
-class OriginQuality(__OriginQuality):
-    """
-    This type contains various attributes commonly used to describe the quality
-    of an origin, e. g., errors, azimuthal coverage, etc. Origin objects have
-    an optional attribute of the type OriginQuality.
-
-    :type associated_phase_count: int, optional
-    :param associated_phase_count: Number of associated phases, regardless of
-        their use for origin computation.
-    :type used_phase_count: int, optional
-    :param used_phase_count: Number of defining phases, i. e., phase
-        observations that were actually used for computing the origin. Note
-        that there may be more than one defining phase per station.
-    :type associated_station_count: int, optional
-    :param associated_station_count: Number of stations at which the event was
-        observed.
-    :type used_station_count: int, optional
-    :param used_station_count: Number of stations from which data was used for
-        origin computation.
-    :type depth_phase_count: int, optional
-    :param depth_phase_count: Number of depth phases (typically pP, sometimes
-        sP) used in depth computation.
-    :type standard_error: float, optional
-    :param standard_error: RMS of the travel time residuals of the arrivals
-        used for the origin computation. Unit: s
-    :type azimuthal_gap: float, optional
-    :param azimuthal_gap: Largest azimuthal gap in station distribution as seen
-        from epicenter. For an illustration of azimuthal gap and secondary
-        azimuthal gap (see below), see Fig. 5 of Bond ́ar et al. (2004).
-        Unit: deg
-    :type secondary_azimuthal_gap: float, optional
-    :param secondary_azimuthal_gap: Secondary azimuthal gap in station
-        distribution, i. e., the largest azimuthal gap a station closes.
-        Unit: deg
-    :type ground_truth_level: str, optional
-    :param ground_truth_level: String describing ground-truth level, e. g. GT0,
-        GT5, etc.
-    :type minimum_distance: float, optional
-    :param minimum_distance: Epicentral distance of station closest to the
-        epicenter.  Unit: deg
-    :type maximum_distance: float, optional
-    :param maximum_distance: Epicentral distance of station farthest from the
-        epicenter.  Unit: deg
-    :type median_distance: float, optional
-    :param median_distance: Median epicentral distance of used stations.
-        Unit: deg
-
-    .. note::
-
-        For handling additional information not covered by the QuakeML
-        standard and how to output it to QuakeML see the
-        :ref:`ObsPy Tutorial <quakeml-extra>`.
-    """
-
-
 __ConfidenceEllipsoid = _event_type_class_factory(
     "__ConfidenceEllipsoid",
     class_attributes=[("semi_major_axis_length", float),
@@ -1534,506 +1089,6 @@ class ConfidenceEllipsoid(__ConfidenceEllipsoid):
         direction of the ellipsoid’s minor axis. Corresponds to Tait-Bryan
         angle θ.
         Unit: deg
-
-    .. note::
-
-        For handling additional information not covered by the QuakeML
-        standard and how to output it to QuakeML see the
-        :ref:`ObsPy Tutorial <quakeml-extra>`.
-    """
-
-
-__OriginUncertainty = _event_type_class_factory(
-    "__OriginUncertainty",
-    class_attributes=[("horizontal_uncertainty", float),
-                      ("min_horizontal_uncertainty", float),
-                      ("max_horizontal_uncertainty", float),
-                      ("azimuth_max_horizontal_uncertainty", float),
-                      ("confidence_ellipsoid", ConfidenceEllipsoid),
-                      ("preferred_description", OriginUncertaintyDescription),
-                      ("confidence_level", float)])
-
-
-class OriginUncertainty(__OriginUncertainty):
-    """
-    This class describes the location uncertainties of an origin.
-
-    The uncertainty can be described either as a simple circular horizontal
-    uncertainty, an uncertainty ellipse according to IMS1.0, or a confidence
-    ellipsoid. If multiple uncertainty models are given, the preferred variant
-    can be specified in the attribute ``preferred_description``.
-
-    :type horizontal_uncertainty: float, optional
-    :param horizontal_uncertainty: Circular confidence region, given by single
-        value of horizontal uncertainty. Unit: m
-    :type min_horizontal_uncertainty: float, optional
-    :param min_horizontal_uncertainty: Semi-minor axis of confidence ellipse.
-        Unit: m
-    :type max_horizontal_uncertainty: float, optional
-    :param max_horizontal_uncertainty: Semi-major axis of confidence ellipse.
-        Unit: m
-    :type azimuth_max_horizontal_uncertainty: float, optional
-    :param azimuth_max_horizontal_uncertainty: Azimuth of major axis of
-        confidence ellipse. Measured clockwise from South-North direction at
-        epicenter. Unit: deg
-    :type confidence_ellipsoid: :class:`~obspy.core.event.ConfidenceEllipsoid`,
-        optional
-    :param confidence_ellipsoid: Confidence ellipsoid
-    :type preferred_description: str, optional
-    :param preferred_description: Preferred uncertainty description. Allowed
-        values are the following:
-
-        * horizontal uncertainty
-        * uncertainty ellipse
-        * confidence ellipsoid
-
-    :type confidence_level: float, optional
-    :param confidence_level: Confidence level of the uncertainty, given in
-        percent.
-
-    .. note::
-
-        For handling additional information not covered by the QuakeML
-        standard and how to output it to QuakeML see the
-        :ref:`ObsPy Tutorial <quakeml-extra>`.
-    """
-
-
-__Origin = _event_type_class_factory(
-    "__Origin",
-    class_attributes=[("resource_id", ResourceIdentifier),
-                      ("time", UTCDateTime, ATTRIBUTE_HAS_ERRORS),
-                      ("longitude", float, ATTRIBUTE_HAS_ERRORS),
-                      ("latitude", float, ATTRIBUTE_HAS_ERRORS),
-                      ("depth", float, ATTRIBUTE_HAS_ERRORS),
-                      ("depth_type", OriginDepthType),
-                      ("time_fixed", bool),
-                      ("epicenter_fixed", bool),
-                      ("reference_system_id", ResourceIdentifier),
-                      ("method_id", ResourceIdentifier),
-                      ("earth_model_id", ResourceIdentifier),
-                      ("quality", OriginQuality),
-                      ("origin_type", OriginType),
-                      ("origin_uncertainty", OriginUncertainty),
-                      ("region", str),
-                      ("evaluation_mode", EvaluationMode),
-                      ("evaluation_status", EvaluationStatus),
-                      ("creation_info", CreationInfo)],
-    class_contains=["comments", "arrivals", "composite_times"])
-
-
-class Origin(__Origin):
-    """
-    This class represents the focal time and geographical location of an
-    earthquake hypocenter, as well as additional meta-information. Origin can
-    have objects of type OriginUncertainty and Arrival as child elements.
-
-    :type resource_id: :class:`~obspy.core.event.ResourceIdentifier`
-    :param resource_id: Resource identifier of Origin.
-    :type force_resource_id: bool, optional
-    :param force_resource_id: If set to False, the automatic initialization of
-        `resource_id` attribute in case it is not specified will be skipped.
-    :type time: :class:`~obspy.core.utcdatetime.UTCDateTime`
-    :param time: Focal time.
-    :type time_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param time_errors: AttribDict containing error quantities.
-    :type longitude: float
-    :param longitude: Hypocenter longitude, with respect to the World Geodetic
-        System 1984 (WGS84) reference system. Unit: deg
-    :type longitude_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param longitude_errors: AttribDict containing error quantities.
-    :type latitude: float
-    :param latitude: Hypocenter latitude, with respect to the WGS84 reference
-        system. Unit: deg
-    :type latitude_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param latitude_errors: AttribDict containing error quantities.
-    :type depth: float, optional
-    :param depth: Depth of hypocenter with respect to the nominal sea level
-        given by the WGS84 geoid. Positive values indicate hypocenters below
-        sea level. For shallow hypocenters, the depth value can be negative.
-        Note: Other standards use different conventions for depth measurement.
-        As an example, GSE2.0, defines depth with respect to the local surface.
-        If event data is converted from other formats to QuakeML, depth values
-        may have to be modified accordingly. Unit: m
-    :type depth_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param depth_errors: AttribDict containing error quantities.
-    :type depth_type: str, optional
-    :param depth_type: Type of depth determination. Allowed values are the
-        following:
-
-        * ``"from location"``
-        * ``"from moment tensor inversion"``
-        * ``"from modeling of broad-band P waveforms"``
-        * ``"constrained by depth phases"``
-        * ``"constrained by direct phases"``
-        * ``"constrained by depth and direct phases"``
-        * ``"operator assigned"``
-        * ``"other"``
-
-    :type time_fixed: bool, optional
-    :param time_fixed: True if focal time was kept fixed for computation of the
-        Origin.
-    :type epicenter_fixed: bool, optional
-    :param epicenter_fixed: True if epicenter was kept fixed for computation of
-        Origin.
-    :type reference_system_id: :class:`~obspy.core.event.ResourceIdentifier`,
-        optional
-    :param reference_system_id: Identifies the reference system used for
-        hypocenter determination. This is only necessary if a modified version
-        of the standard (with local extensions) is used that provides a
-        non-standard coordinate system.
-    :type method_id: :class:`~obspy.core.event.ResourceIdentifier`, optional
-    :param method_id: Identifies the method used for locating the event.
-    :type earth_model_id: :class:`~obspy.core.event.ResourceIdentifier`,
-        optional
-    :param earth_model_id: Identifies the earth model used in method_id.
-    :type arrivals: list of :class:`~obspy.core.event.Arrival`, optional
-    :param arrivals: List of arrivals associated with the origin.
-    :type composite_times: list of :class:`~obspy.core.event.CompositeTime`,
-        optional
-    :param composite_times: Supplementary information on time of rupture start.
-        Complex descriptions of focal times of historic events are possible,
-        see description of the CompositeTime type. Note that even if
-        compositeTime is used, the mandatory time attribute has to be set, too.
-        It has to be set to the single point in time (with uncertainties
-        allowed) that is most characteristic for the event.
-    :type quality: :class:`~obspy.core.event.OriginQuality`, optional
-    :param quality: Additional parameters describing the quality of an Origin
-        determination.
-    :type origin_type: str, optional
-    :param origin_type: Describes the origin type. Allowed values are the
-        following:
-
-        * ``"hypocenter"``
-        * ``"centroid"``
-        * ``"amplitude"``
-        * ``"macroseismic"``
-        * ``"rupture start"``
-        * ``"rupture end"``
-
-    :type origin_uncertainty: :class:`~obspy.core.event.OriginUncertainty`,
-        optional
-    :param origin_uncertainty: Describes the location uncertainties of an
-        origin.
-    :type region: str, optional
-    :param region: Can be used to describe the geographical region of the
-        epicenter location. Useful if an event has multiple origins from
-        different agencies, and these have different region designations. Note
-        that an event-wide region can be defined in the description attribute
-        of an Event object. The user has to take care that this information
-        corresponds to the region attribute of the preferred Origin.
-    :type evaluation_mode: str, optional
-    :param evaluation_mode: Evaluation mode of Origin. Allowed values are the
-        following:
-
-        * ``"manual"``
-        * ``"automatic"``
-
-    :type evaluation_status: str, optional
-    :param evaluation_status: Evaluation status of Origin. Allowed values are
-        the following:
-
-        * ``"preliminary"``
-        * ``"confirmed"``
-        * ``"reviewed"``
-        * ``"final"``
-        * ``"rejected"``
-        * ``"reported"``
-
-    :type comments: list of :class:`~obspy.core.event.Comment`, optional
-    :param comments: Additional comments.
-    :type creation_info: :class:`~obspy.core.event.CreationInfo`, optional
-    :param creation_info: Creation information used to describe author,
-        version, and creation time.
-
-    .. rubric:: Example
-
-    >>> from obspy.core.event import Origin
-    >>> origin = Origin()
-    >>> origin.resource_id = 'smi:ch.ethz.sed/origin/37465'
-    >>> origin.time = UTCDateTime(0)
-    >>> origin.latitude = 12
-    >>> origin.latitude_errors.confidence_level = 95.0
-    >>> origin.longitude = 42
-    >>> origin.depth_type = 'from location'
-    >>> print(origin)  # doctest: +NORMALIZE_WHITESPACE +ELLIPSIS
-    Origin
-        resource_id: ResourceIdentifier(id="smi:ch.ethz.sed/...")
-               time: UTCDateTime(1970, 1, 1, 0, 0)
-          longitude: 42.0
-           latitude: 12.0 [confidence_level=95.0]
-         depth_type: ...'from location'
-
-    .. note::
-
-        For handling additional information not covered by the QuakeML
-        standard and how to output it to QuakeML see the
-        :ref:`ObsPy Tutorial <quakeml-extra>`.
-    """
-
-
-__StationMagnitudeContribution = _event_type_class_factory(
-    "__StationMagnitudeContribution",
-    class_attributes=[("station_magnitude_id", ResourceIdentifier),
-                      ("residual", float),
-                      ("weight", float)])
-
-
-class StationMagnitudeContribution(__StationMagnitudeContribution):
-    """
-    This class describes the weighting of magnitude values from several
-    StationMagnitude objects for computing a network magnitude estimation.
-
-    :type station_magnitude_id: :class:`~obspy.core.event.ResourceIdentifier`,
-        optional
-    :param station_magnitude_id: Refers to the resource_id of a
-        StationMagnitude object.
-    :type residual: float, optional
-    :param residual: Residual of magnitude computation.
-    :type weight: float, optional
-    :param weight: Weight of the magnitude value from class StationMagnitude
-        for computing the magnitude value in class Magnitude. Note that there
-        is no rule for the sum of the weights of all station magnitude
-        contributions to a specific network magnitude. In particular, the
-        weights are not required to sum up to unity.
-
-    .. note::
-
-        For handling additional information not covered by the QuakeML
-        standard and how to output it to QuakeML see the
-        :ref:`ObsPy Tutorial <quakeml-extra>`.
-    """
-
-
-__Magnitude = _event_type_class_factory(
-    "__Magnitude",
-    class_attributes=[("resource_id", ResourceIdentifier),
-                      ("mag", float, ATTRIBUTE_HAS_ERRORS),
-                      ("magnitude_type", str),
-                      ("origin_id", ResourceIdentifier),
-                      ("method_id", ResourceIdentifier),
-                      ("station_count", int),
-                      ("azimuthal_gap", float),
-                      ("evaluation_mode", EvaluationMode),
-                      ("evaluation_status", EvaluationStatus),
-                      ("creation_info", CreationInfo)],
-    class_contains=["comments", "station_magnitude_contributions"])
-
-
-class Magnitude(__Magnitude):
-    """
-    Describes a magnitude which can, but does not need to be associated with an
-    origin.
-
-    Association with an origin is expressed with the optional attribute
-    originID. It is either a combination of different magnitude estimations, or
-    it represents the reported magnitude for the given event.
-
-    :type resource_id: :class:`~obspy.core.event.ResourceIdentifier`
-    :param resource_id: Resource identifier of Magnitude.
-    :type force_resource_id: bool, optional
-    :param force_resource_id: If set to False, the automatic initialization of
-        `resource_id` attribute in case it is not specified will be skipped.
-    :type mag: float
-    :param mag: Resulting magnitude value from combining values of type
-        :class:`~obspy.core.event.StationMagnitude`. If no estimations are
-        available, this value can represent the reported magnitude.
-    :type mag_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param mag_errors: AttribDict containing error quantities.
-    :type magnitude_type: str, optional
-    :param magnitude_type: Describes the type of magnitude. This is a free-text
-        field because it is impossible to cover all existing magnitude type
-        designations with an enumeration. Possible values are:
-
-        * unspecified magnitude (``'M'``),
-        * local magnitude (``'ML'``),
-        * body wave magnitude (``'Mb'``),
-        * surface wave magnitude (``'MS'``),
-        * moment magnitude (``'Mw'``),
-        * duration magnitude (``'Md'``)
-        * coda magnitude (``'Mc'``)
-        * ``'MH'``, ``'Mwp'``, ``'M50'``, ``'M100'``, etc.
-
-    :type origin_id: :class:`~obspy.core.event.ResourceIdentifier`, optional
-    :param origin_id: Reference to an origin’s resource_id if the magnitude has
-        an associated Origin.
-    :type method_id: :class:`~obspy.core.event.ResourceIdentifier`, optional
-    :param method_id: Identifies the method of magnitude estimation. Users
-        should avoid to give contradictory information in method_id and
-        magnitude_type.
-    :type station_count: int, optional
-    :param station_count: Number of used stations for this magnitude
-        computation.
-    :type azimuthal_gap: float, optional
-    :param azimuthal_gap: Azimuthal gap for this magnitude computation.
-        Unit: deg
-    :type evaluation_mode: str, optional
-    :param evaluation_mode: Evaluation mode of Magnitude. Allowed values are
-        the following:
-
-        * ``"manual"``
-        * ``"automatic"``
-
-    :type evaluation_status:
-        :class:`~obspy.core.event_header.EvaluationStatus`, optional
-    :param evaluation_status: Evaluation status of Magnitude. Allowed values
-        are the following:
-
-        * ``"preliminary"``
-        * ``"confirmed"``
-        * ``"reviewed"``
-        * ``"final"``
-        * ``"rejected"``
-        * ``"reported"``
-
-    :type comments: list of :class:`~obspy.core.event.Comment`, optional
-    :param comments: Additional comments.
-    :type station_magnitude_contributions: list of
-        :class:`~obspy.core.event.StationMagnitudeContribution`.
-    :param station_magnitude_contributions: StationMagnitudeContribution
-        instances associated with the Magnitude.
-    :type creation_info: :class:`~obspy.core.event.CreationInfo`, optional
-    :param creation_info: Creation information used to describe author,
-        version, and creation time.
-
-    .. note::
-
-        For handling additional information not covered by the QuakeML
-        standard and how to output it to QuakeML see the
-        :ref:`ObsPy Tutorial <quakeml-extra>`.
-    """
-
-
-__StationMagnitude = _event_type_class_factory(
-    "__StationMagnitude",
-    class_attributes=[("resource_id", ResourceIdentifier),
-                      ("origin_id", ResourceIdentifier),
-                      ("mag", float, ATTRIBUTE_HAS_ERRORS),
-                      ("station_magnitude_type", str),
-                      ("amplitude_id", ResourceIdentifier),
-                      ("method_id", ResourceIdentifier),
-                      ("waveform_id", WaveformStreamID),
-                      ("creation_info", CreationInfo)],
-    class_contains=["comments"])
-
-
-class StationMagnitude(__StationMagnitude):
-    """
-    This class describes the magnitude derived from a single waveform stream.
-
-    :type resource_id: :class:`~obspy.core.event.ResourceIdentifier`, optional
-    :param resource_id: Resource identifier of StationMagnitude.
-    :type force_resource_id: bool, optional
-    :param force_resource_id: If set to False, the automatic initialization of
-        `resource_id` attribute in case it is not specified will be skipped.
-    :type origin_id: :class:`~obspy.core.event.ResourceIdentifier`, optional
-    :param origin_id: Reference to an origin’s ``resource_id`` if the
-        StationMagnitude has an associated :class:`~obspy.core.event.Origin`.
-    :type mag: float
-    :param mag: Estimated magnitude.
-    :type mag_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param mag_errors: AttribDict containing error quantities.
-    :type station_magnitude_type: str, optional
-    :param station_magnitude_type: See :class:`~obspy.core.event.Magnitude`
-    :type amplitude_id: :class:`~obspy.core.event.ResourceIdentifier`, optional
-    :param amplitude_id: Identifies the data source of the StationMagnitude.
-        For magnitudes derived from amplitudes in waveforms (e.g., local
-        magnitude ML), amplitudeID points to publicID in class Amplitude.
-    :type method_id: :class:`~obspy.core.event.ResourceIdentifier`, optional
-    :param method_id: See :class:`~obspy.core.event.Magnitude`
-    :type waveform_id: :class:`~obspy.core.event.WaveformStreamID`, optional
-    :param waveform_id: Identifies the waveform stream. This element can be
-        helpful if no amplitude is referenced, or the amplitude is not
-        available in the context. Otherwise, it would duplicate the waveform_id
-        provided there and can be omitted.
-    :type comments: list of :class:`~obspy.core.event.Comment`, optional
-    :param comments: Additional comments.
-    :type creation_info: :class:`~obspy.core.event.CreationInfo`, optional
-    :param creation_info: Creation information used to describe author,
-        version, and creation time.
-
-    .. note::
-
-        For handling additional information not covered by the QuakeML
-        standard and how to output it to QuakeML see the
-        :ref:`ObsPy Tutorial <quakeml-extra>`.
-    """
-
-
-__EventDescription = _event_type_class_factory(
-    "__EventDescription",
-    class_attributes=[("text", str),
-                      ("type", EventDescriptionType)])
-
-
-class EventDescription(__EventDescription):
-    """
-    Free-form string with additional event description. This can be a
-    well-known name, like 1906 San Francisco Earthquake. A number of categories
-    can be given in type.
-
-    :type text: str, optional
-    :param text: Free-form text with earthquake description.
-    :type type: str, optional
-    :param type: Category of earthquake description. Values
-        can be taken from the following:
-
-        * ``"felt report"``
-        * ``"Flinn-Engdahl region"``
-        * ``"local time"``
-        * ``"tectonic summary"``
-        * ``"nearest cities"``
-        * ``"earthquake name"``
-        * ``"region name"``
-
-    .. note::
-
-        For handling additional information not covered by the QuakeML
-        standard and how to output it to QuakeML see the
-        :ref:`ObsPy Tutorial <quakeml-extra>`.
-    """
-
-
-__Tensor = _event_type_class_factory(
-    "__Tensor",
-    class_attributes=[("m_rr", float, ATTRIBUTE_HAS_ERRORS),
-                      ("m_tt", float, ATTRIBUTE_HAS_ERRORS),
-                      ("m_pp", float, ATTRIBUTE_HAS_ERRORS),
-                      ("m_rt", float, ATTRIBUTE_HAS_ERRORS),
-                      ("m_rp", float, ATTRIBUTE_HAS_ERRORS),
-                      ("m_tp", float, ATTRIBUTE_HAS_ERRORS)])
-
-
-class Tensor(__Tensor):
-    """
-    The Tensor class represents the six moment-tensor elements Mrr, Mtt, Mpp,
-    Mrt, Mrp, Mtp in the spherical coordinate system defined by local upward
-    vertical (r), North-South (t), and West-East (p) directions.
-
-    :type m_rr: float
-    :param m_rr: Moment-tensor element Mrr. Unit: Nm
-    :type m_rr_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param m_rr_errors: AttribDict containing error quantities.
-    :type m_tt: float
-    :param m_tt: Moment-tensor element Mtt. Unit: Nm
-    :type m_tt_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param m_tt_errors: AttribDict containing error quantities.
-    :type m_pp: float
-    :param m_pp: Moment-tensor element Mpp. Unit: Nm
-    :type m_pp_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param m_pp_errors: AttribDict containing error quantities.
-    :type m_rt: float
-    :param m_rt: Moment-tensor element Mrt. Unit: Nm
-    :type m_rt_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param m_rt_errors: AttribDict containing error quantities.
-    :type m_rp: float
-    :param m_rp: Moment-tensor element Mrp. Unit: Nm
-    :type m_rp_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param m_rp_errors: AttribDict containing error quantities.
-    :type m_tp: float
-    :param m_tp: Moment-tensor element Mtp. Unit: Nm
-    :type m_tp_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param m_tp_errors: AttribDict containing error quantities.
 
     .. note::
 
@@ -2078,370 +1133,6 @@ class DataUsed(__DataUsed):
     :param shortest_period: Shortest period present in data. Unit: s
     :type longest_period: float, optional
     :param longest_period: Longest period present in data. Unit: s
-
-    .. note::
-
-        For handling additional information not covered by the QuakeML
-        standard and how to output it to QuakeML see the
-        :ref:`ObsPy Tutorial <quakeml-extra>`.
-    """
-
-
-__SourceTimeFunction = _event_type_class_factory(
-    "__SourceTimeFunction",
-    class_attributes=[("type", SourceTimeFunctionType),
-                      ("duration", float),
-                      ("rise_time", float),
-                      ("decay_time", float)])
-
-
-class SourceTimeFunction(__SourceTimeFunction):
-    """
-    Source time function used in moment-tensor inversion.
-
-    :type type: str
-    :param type: Type of source time function. Values can be taken from the
-        following:
-
-        * ``"box car"``,
-        * ``"triangle"``,
-        * ``"trapezoid"``,
-        * ``"unknown"``
-
-    :type duration: float
-    :param duration: Source time function duration. Unit: s
-    :type rise_time: float, optional
-    :param rise_time: Source time function rise time. Unit: s
-    :type decay_time: float, optional
-    :param decay_time: Source time function decay time. Unit: s
-
-    .. note::
-
-        For handling additional information not covered by the QuakeML
-        standard and how to output it to QuakeML see the
-        :ref:`ObsPy Tutorial <quakeml-extra>`.
-    """
-
-
-__NodalPlane = _event_type_class_factory(
-    "__NodalPlane",
-    class_attributes=[("strike", float, ATTRIBUTE_HAS_ERRORS),
-                      ("dip", float, ATTRIBUTE_HAS_ERRORS),
-                      ("rake", float, ATTRIBUTE_HAS_ERRORS)])
-
-
-class NodalPlane(__NodalPlane):
-    """
-    This class describes a nodal plane using the attributes strike, dip, and
-    rake. For a definition of the angles see Aki & Richards (1980).
-
-    :type strike: float
-    :param strike: Strike angle of nodal plane. Unit: deg
-    :type strike_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param strike_errors: AttribDict containing error quantities.
-    :type dip: float
-    :param dip: Dip angle of nodal plane. Unit: deg
-    :type dip_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param dip_errors: AttribDict containing error quantities.
-    :type rake: float
-    :param rake: Rake angle of nodal plane. Unit: deg
-    :type rake_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param rake_errors: AttribDict containing error quantities.
-
-    .. note::
-
-        For handling additional information not covered by the QuakeML
-        standard and how to output it to QuakeML see the
-        :ref:`ObsPy Tutorial <quakeml-extra>`.
-    """
-
-
-__Axis = _event_type_class_factory(
-    "__Axis",
-    class_attributes=[("azimuth", float, ATTRIBUTE_HAS_ERRORS),
-                      ("plunge", float, ATTRIBUTE_HAS_ERRORS),
-                      ("length", float, ATTRIBUTE_HAS_ERRORS)])
-
-
-class Axis(__Axis):
-    """
-    This class describes an eigenvector of a moment tensor expressed in its
-    principal-axes system. It uses the angles azimuth, plunge, and the
-    eigenvalue length.
-
-    :type azimuth: float
-    :param azimuth: Azimuth of eigenvector of moment tensor expressed in
-        principal-axes system. Measured clockwise from South-North direction at
-        epicenter. Unit: deg
-    :type azimuth_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param azimuth_errors: AttribDict containing error quantities.
-    :type plunge: float
-    :param plunge: Plunge of eigenvector of moment tensor expressed in
-        principal-axes system. Measured against downward vertical direction at
-        epicenter. Unit: deg
-    :type plunge_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param plunge_errors: AttribDict containing error quantities.
-    :type length: float
-    :param length: Eigenvalue of moment tensor expressed in principal-axes
-        system. Unit: Nm
-    :type length_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param length_errors: AttribDict containing error quantities.
-
-    .. note::
-
-        For handling additional information not covered by the QuakeML
-        standard and how to output it to QuakeML see the
-        :ref:`ObsPy Tutorial <quakeml-extra>`.
-    """
-
-
-__NodalPlanes = _event_type_class_factory(
-    "__NodalPlanes",
-    class_attributes=[("nodal_plane_1", NodalPlane),
-                      ("nodal_plane_2", NodalPlane),
-                      ("preferred_plane", int)])
-
-
-class NodalPlanes(__NodalPlanes):
-    """
-    This class describes the nodal planes of a double-couple moment-tensor
-    solution. The attribute ``preferred_plane`` can be used to define which
-    plane is the preferred one.
-
-    :type nodal_plane_1: :class:`~obspy.core.event.NodalPlane`, optional
-    :param nodal_plane_1: First nodal plane of double-couple moment tensor
-        solution.
-    :type nodal_plane_2: :class:`~obspy.core.event.NodalPlane`, optional
-    :param nodal_plane_2: Second nodal plane of double-couple moment tensor
-        solution.
-    :type preferred_plane: int, optional
-    :param preferred_plane: Indicator for preferred nodal plane of moment
-        tensor solution. It can take integer values ``1`` or ``2``.
-
-    .. note::
-
-        For handling additional information not covered by the QuakeML
-        standard and how to output it to QuakeML see the
-        :ref:`ObsPy Tutorial <quakeml-extra>`.
-    """
-
-
-__PrincipalAxes = _event_type_class_factory(
-    "__PrincipalAxes",
-    class_attributes=[("t_axis", Axis),
-                      ("p_axis", Axis),
-                      ("n_axis", Axis)])
-
-
-class PrincipalAxes(__PrincipalAxes):
-    """
-    This class describes the principal axes of a double-couple moment tensor
-    solution. t_axis and p_axis are required, while n_axis is optional.
-
-    :type t_axis: :class:`~obspy.core.event.Axis`
-    :param t_axis: T (tension) axis of a double-couple moment tensor solution.
-    :type p_axis: :class:`~obspy.core.event.Axis`
-    :param p_axis: P (pressure) axis of a double-couple moment tensor solution.
-    :type n_axis: :class:`~obspy.core.event.Axis`, optional
-    :param n_axis: N (neutral) axis of a double-couple moment tensor solution.
-
-    .. note::
-
-        For handling additional information not covered by the QuakeML
-        standard and how to output it to QuakeML see the
-        :ref:`ObsPy Tutorial <quakeml-extra>`.
-    """
-
-
-__MomentTensor = _event_type_class_factory(
-    "__MomentTensor",
-    class_attributes=[("resource_id", ResourceIdentifier),
-                      ("derived_origin_id", ResourceIdentifier),
-                      ("moment_magnitude_id", ResourceIdentifier),
-                      ("scalar_moment", float, ATTRIBUTE_HAS_ERRORS),
-                      ("tensor", Tensor),
-                      ("variance", float),
-                      ("variance_reduction", float),
-                      ("double_couple", float),
-                      ("clvd", float),
-                      ("iso", float),
-                      ("greens_function_id", ResourceIdentifier),
-                      ("filter_id", ResourceIdentifier),
-                      ("source_time_function", SourceTimeFunction),
-                      ("method_id", ResourceIdentifier),
-                      ("category", MomentTensorCategory),
-                      ("inversion_type", MTInversionType),
-                      ("creation_info", CreationInfo)],
-    class_contains=["comments", "data_used"])
-
-
-class MomentTensor(__MomentTensor):
-    """
-    This class represents a moment tensor solution for an event. It is an
-    optional part of a FocalMechanism description.
-
-    :type resource_id: :class:`~obspy.core.event.ResourceIdentifier`
-    :param resource_id: Resource identifier of MomentTensor.
-    :type force_resource_id: bool, optional
-    :param force_resource_id: If set to False, the automatic initialization of
-        `resource_id` attribute in case it is not specified will be skipped.
-    :type derived_origin_id: :class:`~obspy.core.event.ResourceIdentifier`
-    :param derived_origin_id: Refers to the resource_id of the Origin derived
-        in the moment tensor inversion.
-    :type moment_magnitude_id: :class:`~obspy.core.event.ResourceIdentifier`,
-        optional
-    :param moment_magnitude_id: Refers to the publicID of the Magnitude object
-        which represents the derived moment magnitude.
-    :type scalar_moment: float, optional
-    :param scalar_moment: Scalar moment as derived in moment tensor inversion.
-        Unit: Nm
-    :type scalar_moment_errors: :class:`~obspy.core.util.attribdict.AttribDict`
-    :param scalar_moment_errors: AttribDict containing error quantities.
-    :type tensor: :class:`~obspy.core.event.Tensor`, optional
-    :param tensor: Tensor object holding the moment tensor elements.
-    :type variance: float, optional
-    :param variance: Variance of moment tensor inversion.
-    :type variance_reduction: float, optional
-    :param variance_reduction: Variance reduction of moment tensor inversion,
-        given in percent (Dreger 2003). This is a goodness-of-fit measure.
-    :type double_couple: float, optional
-    :param double_couple: Double couple parameter obtained from moment tensor
-        inversion (decimal fraction between 0 and 1).
-    :type clvd: float, optional
-    :param clvd: CLVD (compensated linear vector dipole) parameter obtained
-        from moment tensor inversion (decimal fraction between 0 and 1).
-    :type iso: float, optional
-    :param iso: Isotropic part obtained from moment tensor inversion (decimal
-        fraction between 0 and 1).
-    :type greens_function_id: :class:`~obspy.core.event.ResourceIdentifier`,
-        optional
-    :param greens_function_id: Resource identifier of the Green’s function used
-        in moment tensor inversion.
-    :type filter_id: :class:`~obspy.core.event.ResourceIdentifier`, optional
-    :param filter_id: Resource identifier of the filter setup used in moment
-        tensor inversion.
-    :type source_time_function: :class:`~obspy.core.event.SourceTimeFunction`,
-        optional
-    :param source_time_function: Source time function used in moment-tensor
-        inversion.
-    :type data_used: list of :class:`~obspy.core.event.DataUsed`, optional
-    :param data_used: Describes waveform data used for moment-tensor inversion.
-    :type method_id: :class:`~obspy.core.event.ResourceIdentifier`, optional
-    :param method_id: Resource identifier of the method used for moment-tensor
-        inversion.
-    :type category: str, optional
-    :param category: Moment tensor category. Values can be taken from the
-        following:
-
-        * ``"teleseismic"``,
-        * ``"regional"``
-
-    :type inversion_type: str, optional
-    :param inversion_type: Moment tensor inversion type. Users should avoid to
-        give contradictory information in inversion_type and method_id. Values
-        can be taken from the following:
-
-        * ``"general"``,
-        * ``"zero trace"``,
-        * ``"double couple"``
-
-    :type comments: list of :class:`~obspy.core.event.Comment`, optional
-    :param comments: Additional comments.
-    :type creation_info: :class:`~obspy.core.event.CreationInfo`, optional
-    :param creation_info: Creation information used to describe author,
-        version, and creation time.
-
-    .. note::
-
-        For handling additional information not covered by the QuakeML
-        standard and how to output it to QuakeML see the
-        :ref:`ObsPy Tutorial <quakeml-extra>`.
-    """
-
-
-__FocalMechanism = _event_type_class_factory(
-    "__FocalMechanism",
-    class_attributes=[("resource_id", ResourceIdentifier),
-                      ("triggering_origin_id", ResourceIdentifier),
-                      ("nodal_planes", NodalPlanes),
-                      ("principal_axes", PrincipalAxes),
-                      ("azimuthal_gap", float),
-                      ("station_polarity_count", int),
-                      ("misfit", float),
-                      ("station_distribution_ratio", float),
-                      ("method_id", ResourceIdentifier),
-                      ("evaluation_mode", EvaluationMode),
-                      ("evaluation_status", EvaluationStatus),
-                      ("moment_tensor", MomentTensor),
-                      ("creation_info", CreationInfo)],
-    class_contains=['waveform_id', 'comments'])
-
-
-class FocalMechanism(__FocalMechanism):
-    """
-    This class describes the focal mechanism of an event. It includes different
-    descriptions like nodal planes, principal axes, and a moment tensor. The
-    moment tensor description is provided by objects of the class MomentTensor
-    which can be specified as child elements of FocalMechanism.
-
-    :type resource_id: :class:`~obspy.core.event.ResourceIdentifier`
-    :param resource_id: Resource identifier of FocalMechanism.
-    :type force_resource_id: bool, optional
-    :param force_resource_id: If set to False, the automatic initialization of
-        `resource_id` attribute in case it is not specified will be skipped.
-    :type triggering_origin_id: :class:`~obspy.core.event.ResourceIdentifier`,
-        optional
-    :param triggering_origin_id: Refers to the resource_id of the triggering
-        origin.
-    :type nodal_planes: :class:`~obspy.core.event.NodalPlanes`, optional
-    :param nodal_planes: Nodal planes of the focal mechanism.
-    :type principal_axes: :class:`~obspy.core.event.PrincipalAxes`, optional
-    :param principal_axes: Principal axes of the focal mechanism.
-    :type azimuthal_gap: float, optional
-    :param azimuthal_gap: Largest azimuthal gap in distribution of stations
-        used for determination of focal mechanism. Unit: deg
-    :type station_polarity_count: int, optional
-    :param station_polarity_count:
-    :type misfit: float, optional
-    :param misfit: Fraction of misfit polarities in a first-motion focal
-        mechanism determination. Decimal fraction between 0 and 1.
-    :type station_distribution_ratio: float, optional
-    :param station_distribution_ratio: Station distribution ratio (STDR)
-        parameter. Indicates how the stations are distributed about the focal
-        sphere (Reasenberg and Oppenheimer 1985). Decimal fraction between 0
-        and 1.
-    :type method_id: :class:`~obspy.core.event.ResourceIdentifier`, optional
-    :param method_id: Resource identifier of the method used for determination
-        of the focal mechanism.
-    :type waveform_id: list of :class:`~obspy.core.event.WaveformStreamID`,
-        optional
-    :param waveform_id: Refers to a set of waveform streams from which the
-        focal mechanism was derived.
-    :type evaluation_mode: str, optional
-    :param evaluation_mode: Evaluation mode of FocalMechanism. Allowed values
-        are the following:
-
-        * ``"manual"``
-        * ``"automatic"``
-
-    :type evaluation_status: str, optional
-    :param evaluation_status: Evaluation status of FocalMechanism. Allowed
-        values are the following:
-
-        * ``"preliminary"``
-        * ``"confirmed"``
-        * ``"reviewed"``
-        * ``"final"``
-        * ``"rejected"``
-        * ``"reported"``
-
-    :type moment_tensor: :class:`~obspy.core.event.MomentTensor`, optional
-    :param moment_tensor: Moment tensor description for this focal mechanism.
-    :type comments: list of :class:`~obspy.core.event.Comment`, optional
-    :param comments: Additional comments.
-    :type creation_info: :class:`~obspy.core.event.CreationInfo`, optional
-    :param creation_info: Creation information used to describe author,
-        version, and creation time.
 
     .. note::
